+++ recipes/upgrade.rb
@@ -23,8 +23,8 @@ unless node['splunk']['upgrade_enabled']
   raise 'Failed to upgrade'
 end
 
-splunk_package = is_server? ? 'splunk' : 'splunkforwarder'
-url_type = is_server? ? 'server' : 'forwarder'
+splunk_package = server? ? 'splunk' : 'splunkforwarder'
+url_type = server? ? 'server' : 'forwarder'
 
 splunk_installer "#{splunk_package} upgrade" do
   action :upgrade